refactor: extract the handshake INSTRUCTIONS constant into instructions.py
Moves the INSTRUCTIONS constant (server.py 94-319, the text handed to
MCPServer(..., instructions=...) at the client handshake) into a new
leaf module, src/comfy_mcp/instructions.py, byte-for-byte. server.py
reaches it module-qualified (instructions.INSTRUCTIONS) and re-exports
nothing, per the repo's no-re-export convention (AGENTS.md ~117-121):
a test patching a moved name on server would otherwise silently patch
a name nothing reads. Six test files that read server.INSTRUCTIONS
(test_freshness_docs, test_install_node, test_payload_budget,
test_routing_instructions, test_subgraph_templates, test_switch_version)
now import instructions and read instructions.INSTRUCTIONS instead;
none of them patch it, so none needed anything more. AGENTS.md's module
table gains an instructions.py row.
